fix: duplicate dependency on a nested project would fail check (#89)
Projects that were added to the project via `implementation project(path: ":features:discover")`
would fail to be correctly checked as a duplicate dependency and would always be flagged as true, 
due to the parsing only checking the first part of the path and not the whole thing. Update the
Dependency model to have a method to check which is a bit smarter. As part of this cleaned up some
of the logic in a few areas.

Closes #88