chore: remove novodo plugin (#94)
* chore: remove the old novodo plugin and use the bintray plugins

Use the base plugins to upload builds to Bintray instead of relying on the third party novodo plugin
which wasn't working.

* ci: update github actions to run install before deploy

Install must be run before deploying to bintray

* chore: move install into bintray gradle file

Instead of having duplicate gradle files, combine them into one since they shouldn't need to be
edited. Also fixed a few issues that occured along the way

* chore: change imports around to be lexicographic

* chore: remove the warnings for not supplying gradle properties

Remove the warnings as they were a bit annoying to have in CI. Instead will add to the documentation
that they are required

* chore: add new task to deploy to bintray

Add task 'deploy' that will verify that the correct params are passed in. It will then handle
running install and bintrayUpload afterwards

* ci: update ci script to use the new deploy gradle task

Remove the need to call install and bintray when we can just use the new deploy task, which will
also ensure that the correct params are passed in